[add] i18n.js
